@@ -6,24 +6,40 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=3.7.0 <4.0.0"
 
 dependencies:
-  async: ^2.13.0
-  fake_async: ^1.3.3
-  material_color_utilities: ^0.11.1  # Compatible avec votre version de Flutter
-  vector_math: 2.1.4
-  google_fonts: ^6.1.0
   flutter:
     sdk: flutter
+
+  cupertino_icons: ^1.0.2
+
+  google_fonts: ^6.1.0
   http: ^1.3.0
   audioplayers: ^6.4.0
   shared_preferences: ^2.5.3
 
-dev_dependencies:
-  flutter_lints: ^5.0.0  # Version stable compatible avec Flutter 3.29.1
+  provider: ^6.1.2
+  intl: ^0.19.0 # Assure-toi que cette version ou une plus récente est compatible
 
+  # Dépendances pour le cache offline (Hive)
+  hive: ^2.2.3 # Vérifie la dernière version sur pub.dev
+  hive_flutter: ^1.1.0 # Vérifie la dernière version sur pub.dev
+  path_provider: ^2.0.0 # Pour trouver le chemin où stocker les fichiers Hive
+
+  # Optionnel: Pour vérifier la connexion internet (utile pour la stratégie de cache)
+  # connectivity_plus: ^6.0.3 # Vérifie la dernière version
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+
+  flutter_lints: ^3.0.0
+
+  # Dépendances de développement pour Hive Generator
+  hive_generator: ^2.0.1 # Vérifie la dernière version
+  build_runner: ^2.4.9 # Vérifie la dernière version (outil qui exécute hive_generator)
 
 flutter:
   uses-material-design: true
